Simplify styled components into fewer files and rename stuff
