--- packages/website/src/components/coolbar.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const CoolBar = styled.div`
-  min-height: 100%;
-  width: calc(var(--space-single)*.4);
-  background-color: var(--color-dark-blue);
-  flex-shrink: 0;
-`;
-
-export default CoolBar;
