+++ packages/website/src/components/image.js
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
 
-const ImageContainer = styled.div`
+const Image = styled.div`
   height: 200px; 
   justify-content: center;
   display: flex;
   align-items: center;
 `;
 
-export default ImageContainer;
+export default Image;
